[!] zephyr: fix unaligned store exception by making some structures 8-bytes aligned

+++ zephyr/include/drivers/console/uart_console.h
@@ -27,7 +27,7 @@ extern "C" {
 
 #define MAX_LINE_LEN 256
 struct uart_console_input {
-	int _unused;
+	int64_t _unused;
 	char line[MAX_LINE_LEN];
 };
